# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Abuzar Next API
  version: 0.1.0
  description: Tenant-scoped central API for online and synchronized branch operations.
servers:
  - url: https://api.example.invalid
paths:
...
  /v1/transactions/sales:
    post:
      operationId: createSale
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncEnvelope' }
      responses:
        '202':
          description: Sale accepted or deduplicated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: { type: boolean }
                  duplicate: { type: boolean }
                  eventId: { type: string, format: uuid }
                  aggregateId: { type: string, format: uuid }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/transactions/returns:
    post:
      operationId: createReturn
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Return accepted or deduplicated
          content:
            application/json: { schema: { $ref: '#/components/schemas/TransactionResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/transactions/receiving:
    post:
      operationId: createReceiving
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Receiving accepted or deduplicated
          content:
            application/json: { schema: { $ref: '#/components/schemas/TransactionResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/transactions/inventory:
    post:
      operationId: createInventoryTransaction
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Inventory event accepted or deduplicated
          content:
            application/json: { schema: { $ref: '#/components/schemas/TransactionResult' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/shifts/open:
    post:
      operationId: openShift
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: string }
      responses:
        '201': { description: Shift opened }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/shifts/{id}/close:
    post:
      operationId: closeShift
      security: [{ sessionCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: string }
      responses:
        '200': { description: Shift closed }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/sync/push:
    post:
      operationId: pushSync
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events: { type: array, minItems: 1, maxItems: 500, items: { $ref: '#/components/schemas/SyncEnvelope' } }
      responses:
        '202':
          description: Batch result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncBatchResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/sync/pull:
    get:
      operationId: pullSync
      security: [{ sessionCookie: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: Server-authoritative event page
          content:
            application/json:
              schema:
                type: object
                required: [events, nextCursor]
                properties:
                  events: { type: array, items: { $ref: '#/components/schemas/SyncEnvelope' } }
                  nextCursor: { type: integer, format: int64 }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: abuzar_session
  responses:
    Unauthorized:
      description: Authentication required
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
    Conflict:
      description: Idempotency conflict
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
...
  schemas:
...
    SyncEnvelope:
      type: object
      required: [eventId, aggregate, aggregateId, tenantId, branchId, counterId, operatorId, occurredAt, idempotencyKey, schemaVersion, payload]
      properties:
        eventId: { type: string, format: uuid }
        aggregate: { type: string }
        aggregateId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        branchId: { type: string, format: uuid }
        counterId: { type: string, format: uuid }
        operatorId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        idempotencyKey: { type: string }
        schemaVersion: { type: integer, minimum: 1 }
        payload: { type: object, additionalProperties: true }
    SyncBatchResult:
      type: object
      required: [accepted, duplicates, conflicts]
      properties:
        accepted: { type: integer, minimum: 0 }
        duplicates: { type: integer, minimum: 0 }
        conflicts: { type: integer, minimum: 0 }
    TransactionResult:
      type: object
      required: [accepted, duplicate, eventId, aggregateId]
      properties:
        accepted: { type: boolean }
        duplicate: { type: boolean }
        eventId: { type: string, format: uuid }
        aggregateId: { type: string, format: uuid }
...
    ProblemDetails:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }